Make `duration` field as long.
- AudioMessageContent, VideoMessageContent, AudioMessage

+++ line-bot-model/src/main/java/com/linecorp/bot/model/event/message/AudioMessageContent.java
@@ -38,5 +38,5 @@ public class AudioMessageContent implements MessageContent {
 
     String id;
     ContentProvider contentProvider;
-    Integer duration;
+    Long duration;
 }
